requirements.txt
google-cloud-firestore
requests
websocket-client